Fixed a bug on currency exchange

+++ RemoteStocking/Server/Currency.cs
@@ -18,19 +18,19 @@ namespace Server
         public double AWG { get; set; }
         public double AZN { get; set; }
         public double BAM { get; set; }
-        public int BBD { get; set; }
+        public double BBD { get; set; }
         public double BDT { get; set; }
         public double BGN { get; set; }
         public double BHD { get; set; }
-        public int BIF { get; set; }
-        public int BMD { get; set; }
+        public double BIF { get; set; }
+        public double BMD { get; set; }
         public double BND { get; set; }
         public double BOB { get; set; }
         public double BRL { get; set; }
-        public int BSD { get; set; }
+        public double BSD { get; set; }
         public double BTN { get; set; }
         public double BWP { get; set; }
-        public int BYR { get; set; }
+        public double BYR { get; set; }
         public double BZD { get; set; }
         public double CAD { get; set; }
         public double CDF { get; set; }
@@ -40,13 +40,13 @@ namespace Server
         public double CNH { get; set; }
         public double CNY { get; set; }
         public double COP { get; set; }
-        public int CRC { get; set; }
-        public int CUP { get; set; }
+        public double CRC { get; set; }
+        public double CUP { get; set; }
         public double CVE { get; set; }
         public double CZK { get; set; }
         public double DJF { get; set; }
         public double DKK { get; set; }
-        public int DOP { get; set; }
+        public double DOP { get; set; }
         public double DZD { get; set; }
         public double EGP { get; set; }
         public double ETB { get; set; }
@@ -66,27 +66,27 @@ namespace Server
         public double HRK { get; set; }
         public double HTG { get; set; }
         public double HUF { get; set; }
-        public int IDR { get; set; }
+        public double IDR { get; set; }
         public double IEP { get; set; }
         public double ILS { get; set; }
         public double INR { get; set; }
         public double IQD { get; set; }
-        public int IRR { get; set; }
+        public double IRR { get; set; }
         public double ISK { get; set; }
         public double JMD { get; set; }
         public double JOD { get; set; }
         public double JPY { get; set; }
         public double KES { get; set; }
         public double KGS { get; set; }
-        public int KHR { get; set; }
+        public double KHR { get; set; }
         public double KMF { get; set; }
-        public int KPW { get; set; }
+        public double KPW { get; set; }
         public double KRW { get; set; }
         public double KWD { get; set; }
         public double KZT { get; set; }
         public double LAK { get; set; }
         public double LBP { get; set; }
-        public int LKR { get; set; }
+        public double LKR { get; set; }
         public double LRD { get; set; }
         public double LSL { get; set; }
         public double LTL { get; set; }
@@ -94,7 +94,7 @@ namespace Server
         public double LYD { get; set; }
         public double MAD { get; set; }
         public double MDL { get; set; }
-        public int MGA { get; set; }
+        public double MGA { get; set; }
         public double MKD { get; set; }
         public double MMK { get; set; }
         public double MNT { get; set; }
@@ -113,13 +113,13 @@ namespace Server
         public double NPR { get; set; }
         public double NZD { get; set; }
         public double OMR { get; set; }
-        public int PAB { get; set; }
+        public double PAB { get; set; }
         public double PEN { get; set; }
         public double PGK { get; set; }
         public double PHP { get; set; }
         public double PKR { get; set; }
         public double PLN { get; set; }
-        public int PYG { get; set; }
+        public double PYG { get; set; }
         public double QAR { get; set; }
         public double RON { get; set; }
         public double RSD { get; set; }
@@ -132,10 +132,10 @@ namespace Server
         public double SEK { get; set; }
         public double SGD { get; set; }
         public double SHP { get; set; }
-        public int SLL { get; set; }
-        public int SOS { get; set; }
+        public double SLL { get; set; }
+        public double SOS { get; set; }
         public double SRD { get; set; }
-        public int STD { get; set; }
+        public double STD { get; set; }
         public double SVC { get; set; }
         public double SYP { get; set; }
         public double SZL { get; set; }
@@ -149,12 +149,12 @@ namespace Server
         public double TWD { get; set; }
         public double TZS { get; set; }
         public double UAH { get; set; }
-        public int UGX { get; set; }
-        public int USD { get; set; }
+        public double UGX { get; set; }
+        public double USD { get; set; }
         public double UYU { get; set; }
         public double UZS { get; set; }
         public double VEF { get; set; }
-        public int VND { get; set; }
+        public double VND { get; set; }
         public double VUV { get; set; }
         public double WST { get; set; }
         public double XAF { get; set; }
@@ -164,7 +164,7 @@ namespace Server
         public double XPF { get; set; }
         public double YER { get; set; }
         public double ZAR { get; set; }
-        public int ZMK { get; set; }
+        public double ZMK { get; set; }
         public double ZWL { get; set; }
     }
